// frontend/src/sections/settings/devices/labelPrinter.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type UpdateLabelPrinterSettingsVariables = {
  input: {
    address: string;
    labelHeight: number;
    labelWidth: number;
    port: number;
  };
};

export type UpdateLabelPrinterSettingsResult = {
  updateLabelPrinterSettings: ({
  __typename: "LabelPrinterUpdateResult";
} & {
  success: boolean;
}) | ({
  __typename: "UpdateLabelPrinterSettingsError";
} & {
  error: string;
});
};

export const UpdateLabelPrinterSettings = {
  query: "mutation updateLabelPrinterSettings($input: LabelPrinterSettingsInput!) {\n  updateLabelPrinterSettings(input: $input) {\n    ... on LabelPrinterUpdateResult {\n      __typename\n      success\n    }\n    ... on UpdateLabelPrinterSettingsError {\n      __typename\n      error\n    }\n  }\n}",
} as TypedDocument<UpdateLabelPrinterSettingsResult, UpdateLabelPrinterSettingsVariables>;